renaming and removal of linter warnings
