+++ go.mod
@@ -3,9 +3,12 @@ module gotests/amass-engine/events_scheduler
 go 1.21.1
 
 require (
+	github.com/caffix/queue v0.1.5 // direct
+	github.com/google/uuid v1.3.1 // direct
+)
+
+require (
 	golang.org/x/net v0.17.0 // indirect
 	golang.org/x/sys v0.13.0 // indirect
-	github.com/caffix/queue v0.1.4 // direct
 	google.golang.org/protobuf v1.31.0 // indirect
-	github.com/google/uuid v1.3.1 // direct
 )
